requirements.txt
fastapi
jinja2
uvicorn
openai
python-dotenv
python-multipart